% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateMWAD.R
\name{CalculateMWAD}
\alias{CalculateMWAD}
\title{Mean weighted absolute distance for data columns}
\usage{
CalculateMWAD(data, lossRatioColumns, weights = NULL)
}
\arguments{
\item{data}{a data.table}

\item{lossRatioColumns}{a \code{character} vector with loss ratio columns}

\item{weights}{a \code{character} specifying column name with weights. Default is NULL.}
}
\value{
data.table
}
\description{
This function returns MWAD for all columns in \code{data} specified by
  \code{lossRatioColumns}.
}
\examples{
data <- data.table::data.table(col1 = seq(0.8, 1.2, 0.1), col2 = rnorm(5, 1, 0.1),
                               ekspo = runif(5))
res <- firstPack::CalculateMWAD(data, lossRatioColumns = c("col1", "col2"), weights = "ekspo")
}
\seealso{
\code{\link[secondPack]{MWAD}}
}
