% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doubleNum.R
\name{doubleNum}
\alias{doubleNum}
\title{first function
to double the number}
\usage{
doubleNum(x)
}
\arguments{
\item{x}{is a number}
}
\description{
first function
to double the number
}
\examples{
doubleNum(x)
=9
}
\author{
lancezlin
}

