% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{dtSUM}
\alias{dtSUM}
\title{Sum of two columns}
\usage{
dtSUM(data, col1, col2)
}
\arguments{
\item{data}{a data.table}

\item{col1}{first col}

\item{col2}{second col}
}
\value{
data.table
}
\description{
Summing two columns of data
}
