% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simBrM.R
\name{simBrM}
\alias{simBrM}
\title{simBrM}
\usage{
simBrM(n, sigma = 1)
}
\arguments{
\item{n}{number of steps}

\item{sigma}{SD of step size}
}
\value{
an nx2 matrix of positions
}
\description{
Provides a funtions to generte 2d Brownian motions
}
\examples{
x <- simBrM(1000)
}
